Move RememberMe classes in their own folder
