+++ config/Security/RememberMe/RememberMeToken.php
@@ -1,7 +1,8 @@
 <?php
-namespace Config\Security;
+namespace Config\Security\RememberMe;
 
 use App\Model\User;
+use Config\Security\AbstractToken;
 
 class RememberMeToken extends AbstractToken
 {
